==> 1_single-project-exploration.Rmd <==
---
title: "Los Amigos Camera Trapping Exploration"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r wd, include=FALSE}
# DO NOT RUN THIS IF YOU ARE RUNNING LINE BY LINE - ONLY FOR FINAL REPORT (KNIT)
# Set the WD to the project folder (markdowns working directory when you knit the HTML is where the .rmd is located)
require("knitr")
opts_knit$set(root.dir = '../')

```

```{r non-adjustable options, echo=F, include=F}
#Load Packages
list.of.packages <- c("leaflet", "dplyr", "colortools", "kriging", "corrplot", "lubridate", "kableExtra", "rredlist","sf", "usedist", "ggplot2", "ggpubr", "googledrive", "purrr", "plotly", "googlesheets4")

# Check you have them and load them
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)

lapply(list.of.packages, require, character.only = TRUE)
```

```{r googledrive setup and file import, include=FALSE, eval=F}
#1

# run this line, and enter "Yes". It will open up a browser window, and you'll have to allow access to your Google Drive account

x <- drive_find(n_max = 30)
1
# Create the file
dir.create("data")
dir.create("data/raw-data")
dir.create("data/processed-data")


# Wildlife Insights data
# you must have stored you wildlife insights data download packet on google drive - link to it with the URL
folder_url <- "https://drive.google.com/drive/folders/1iFtdGqWDDZw_9JhlgFplhE6Hq9TO0zad"
folder <- drive_get(as_id(folder_url))
csv_files <- drive_ls(folder, type = "csv", recursive=F)

# Clear the folder (overwriting old files)
do.call(file.remove, list(list.files("data/raw-data", full.names = TRUE)))

# Downloads all the files into a subfolder
for (i in 1:nrow(csv_files)) {
    drive_download(as_id(csv_files$id[i]), 
                   path = paste("data/raw-data", csv_files$name[i], sep = "/"), 
                   overwrite=T)
}

# Data entry - target the data entry sheet
folder_url <- "https://drive.google.com/drive/u/0/folders/1nxFVm2Aib1m7TSNEkJjozP-x_GJtM1fo"
folder <- drive_get(as_id(folder_url))
dataentry_files <- drive_ls(folder, recursive=F)

tmp <- dataentry_files[dataentry_files$name %in% c("Entrada_data","Camera_locations_master"),]

# Downloads all the files into a subfolder
for (i in 1:nrow(tmp)) {
    drive_download(as_id(tmp$id[i]), type = "csv",
                   path = paste("data/raw-data", tmp$name[i], sep = "/"), 
                   overwrite=T)
}


```

```{r data setup, include=FALSE}
# Load your data 
dat <- read.csv("data/raw-data/images.csv", header=T)
dep <- read.csv("data/raw-data/deployments.csv", header=T)
pro <- read.csv("data/raw-data/projects.csv", header=T, sep=",")

# EXTRA FOR THIS PROJECT
all.deps <- read.csv("data/raw-data/entrada_data.csv", header=T, sep=",")
colnames(all.deps)[colnames(all.deps)=="deployment_id.LOCATION_CAMERA_STARTDATE"] <- "deployment_id"

#####################################
# Read in speed data
library(googlesheets4)
library(tidyr)
googlesheets4::gs4_auth()
1
1


speed <- read_sheet(drive_get("https://docs.google.com/spreadsheets/d/1-2HdDFhLWqQE-vud_E4klfQzVvpEc56AboXIKZlYN_I"), col_types="cccccttcnc")
head(speed)


# Read in vulture data

carcass <- read_sheet(drive_get("https://docs.google.com/spreadsheets/d/1uQCJFlRHXTbn2Q2pfKMwapMPmy_qH6U-wP0gREDLmZI"), col_types="c")
# Add date time
carcass$Date_time_deployed <- strptime(carcass$Date_time_deployed," %d/%m/%Y %H:%M:%S", tz="America/Lima")
###############################



# Add location ID column
all.deps$deployment_location_id <- substr(all.deps$deployment_id,1,6)


# Camera locations master
all.locs <- read.csv("data/raw-data/Camera_locations_master.csv", header=T, sep=",")

# Remove blank rows
all.deps <- all.deps[all.deps$deployment_id!="__",]

# Remove the coordinate information and left_join with the locations master
all.deps$Latitude <- NULL; all.deps$Longitude <- NULL
head(all.deps)
head(all.locs)

all.deps <- left_join(all.deps, all.locs[,c("Location.Name", "Latitude", "Longitude")], by=c("deployment_location_id"="Location.Name"))

head(dep)
# Timezone  
# They are mainly Reconyx, so use UTC
tz <- "PET"

# Use the strata category to colour plots 
category <- "subproject_name"

# Define a colour from the R options to base the colourscheme
colour <- "lightseagreen"

# Format the date objects
dat$timestamp <- strptime(dat$timestamp, "%Y-%m-%d %H:%M:%S", tz="America/Lima")
dep$start_date <- strptime(dep$start_date, "%Y-%m-%d", tz="UTC")
dep$end_date <- strptime(dep$end_date, "%Y-%m-%d", tz="UTC")

#Calculate days active
dep$days <- dep$end_date-dep$start_date

# Add a deployment_location_id column
dat$deployment_location_id <- substr(dat$deployment_id,1,6)
dep$deployment_location_id <- substr(dep$deployment_id,1,6)

all.deps$subproject_nam <- NA 
all.deps$subproject_nam[substr(all.deps$deployment_id,1,3)=="LAC"] <- "Grids"
all.deps$subproject_nam[substr(all.deps$deployment_id,1,3)=="FEA"] <- "Features"
all.deps$subproject_nam[substr(all.deps$deployment_id,1,3)=="VUL"] <- "Carcass_watching"

# Create a list of currently active deployments in odr of time since deployment
active.deps <- all.deps[all.deps$End.date=="" & is.na(all.deps$End.date)==F,]
active.deps$Start.date <- strptime(active.deps$Start.date, "%d/%m/%y", tz="UTC")
active.deps$Days <- round(as.numeric(strptime(Sys.Date(), "%Y-%m-%d") -  active.deps$Start.date), 1) 

# ADD DEPLOYMENT COVERAGE
tmp <- st_as_sf(all.deps[is.na(all.deps$Longitude)==F,],coords=c("Longitude", "Latitude"),crs=4326)

active.deps[substr(active.deps$deployment_id,1,3)=="FEA",]

```

The sub-projects included in this script are **`r unique(dep$subproject_name)`**.

# Active deployment summary
The project currently has `r nrow(active.deps)` cameras active in the field. The project currently spans `r paste(round(as.numeric(st_area(st_as_sfc(st_bbox(tmp)))/1e+6),1), "km^2")`

## Deployment locations [Active]

```{r, echo=F}
# Generate colours to display the category levels - R needs them as a factor
active.deps <- active.deps[is.na(active.deps$Latitude)==F,]
# Create a continuous palette function
pal <- colorNumeric(
  palette = "Reds",
  domain = active.deps$Days)

active.deps$Cols <- pal(active.deps$Days)

# Make a legend key
tmp<- data.frame(Cols=c(active.deps$Cols[active.deps$Days==min(active.deps$Days)][1],
                  active.deps$Cols[nrow(active.deps)/2],
                  active.deps$Cols[active.deps$Days==max(active.deps$Days)][1]),
           Labels=c(active.deps$Days[active.deps$Days==min(active.deps$Days)][1],
                  active.deps$Days[nrow(active.deps)/2],
                  active.deps$Days[active.deps$Days==max(active.deps$Days)][1]))


```

```{r map active, echo=F}

# Count the number of camera ststions
n.stat <- length(unique(active.deps$deployment_id))

m <- leaflet() %>%
  addProviderTiles(providers$Esri.WorldTopoMap, group="Base") %>%
  addProviderTiles(providers$Esri.WorldImagery, group="Satellite") %>%  # Add satellite data
  
  addCircleMarkers(lng= active.deps$Longitude, lat= active.deps$Latitude,
                   color=active.deps$Cols ,
                   popup=paste( active.deps$deployment_location_id, "-" , active.deps$Days, "days active"), radius=3) %>%
 
  addLegend("topleft", colors = tmp$Cols,  labels = tmp$Labels,
    title = "Days active",
    labFormat = labelFormat(prefix = "$"),
    opacity = 1
  ) %>%
  # Layers control
  addLayersControl(
    baseGroups = c("Satellite", "Base"),
    options = layersControlOptions(collapsed = FALSE)
  )
m

```

### Active locations by subproject

```{r active summaries, echo=F}
tmp <- as.data.frame(table(active.deps$subproject_nam))

p1 <- plot_ly(
  x=tmp$Var1,
  y=tmp$Freq,
  name = "Active deployments",
  type = "bar"
) %>%
  layout(plot_bgcolor = "#f5f5f5", 
         xaxis = list(title = 'Sub-project'), 
         yaxis = list(title = 'Active deployments'),
         title="Active deployments")

tmp <- as.data.frame(table(dep$subproject_name))

p1

```

### List of active deployment locations

In order to help decide the next cameras to check - we also list the currently active cameras below (with the longest running first):

```{r, echo=F}
tmp <- active.deps[, c("deployment_id", "deployment_location_id", "Days")]
tmp <- tmp[order(tmp$Days, decreasing=T),]

tmp <- left_join(tmp, all.locs, by=c("deployment_location_id"= "Location.Name"))

tmp[,c("deployment_id", "deployment_location_id",  "Days", "Project" ,"Group")] %>%
  kbl() %>%
  kable_styling(full_width = T) %>%
  column_spec(1, bold = T, border_right = T )%>%
  scroll_box( height = "200px")

```

# All deployments
To date we have deployed cameras in `r length(unique(all.deps$deployment_location_id))` locations. 

```{r, echo=F, include=F}
# Generate colours to display the category levels - R needs them as a factor
table(all.deps$Subprojet)
all.deps[,category] <- factor(all.deps[,"Subprojet"])
col.cat <- wheel(colour, num = length(levels(all.deps[,category])))
all.deps$Cols <- col.cat[all.deps[,category]]
```

```{r map, echo=F, warning=F, message=F}

# Count the number of camera ststions
n.stat <- length(unique(all.deps$deployment_id))

m <- leaflet() %>%
  addProviderTiles(providers$Esri.WorldTopoMap, group="Base") %>%
  addProviderTiles(providers$Esri.WorldImagery, group="Satellite") %>%  # Add satellite data
  
  addCircleMarkers(lng= all.deps$Longitude, lat= all.deps$Latitude,
                   color= all.deps$Cols,
                   popup=paste( all.deps$Location,  all.deps[,category]), radius=3) %>%
 addLegend("topleft", colors = col.cat,  labels = levels( all.deps[,category]),
    title = category,
    labFormat = labelFormat(prefix = "$"),
    opacity = 1
  ) %>%
  # Layers control
  addLayersControl(
    baseGroups = c("Satellite", "Base"),
    options = layersControlOptions(collapsed = FALSE)
  )
m

```


# Wildlife Insights Summary

```{r, echo=F}
tmp <- as.data.frame(table(dat$identified_by=="Computer vision"))
tmp$Type <- "Labelled"
tmp$Type[tmp$Var1==TRUE] <- "To label"

```

To date, `r Sys.Date()`, we have `r nrow(dat)` images in Wildlife Insights, of which `r round((tmp$Freq[tmp$Type=="Labelled"]/sum(tmp$Freq))*100, 1)`% are identified (labelled) by people:

```{r, echo=F}
par(mfrow=c(1,1))

f1 <- plot_ly(tmp,y=~tmp$Freq, x=tmp$Type,  type="bar", name=tmp$Type) %>% 
        layout(yaxis = list(title = 'Images'), barmode = 'group')

f1
```

The breakdown by labeler is as follows:

```{r, echo=F}
tmp2 <- as.data.frame(table(dat$identified_by[!dat$identified_by%in% c("Computer vision", "")]))

colnames(tmp2)[1] <- "Person"

f2 <- plot_ly(tmp2,y=~tmp2$Freq, x=tmp2$Person,  type="bar", name=tmp2$Person) %>% 
        layout(yaxis = list(title = 'Images labelled'), barmode = 'group')


f2

```

## Species labels

```{r, echo=F}
# Reduce to verfied data
vdat <- dat[!dat$identified_by %in% c("Computer vision", ""),]
# Remove blanks
vdat$sp <- paste(vdat$genus, vdat$species)
vdat <- vdat[vdat$is_blank==0,]
vdat <- vdat[vdat$species!="No CV Result",]
vdat <- vdat[vdat$common_name!="Blank",]

# Do all species with a latin name have a common name - YES use common name
#as.data.frame(table(paste(vdat$sp, vdat$common_name)))

# if the species name is blank it must be to family, order or class level
```

To date we have `r nrow(vdat)` image labels verified by human observers (after removing `r nrow(dat[dat$is_blank==1 | dat$species=="No CV Result" | dat$common_name =="Blank",])` blanks). The species categories are as follows:

```{r, include=F}

# OLD VERSION
n.loc <- length(unique(dep$deployment_location_id))

layout(matrix(c(1,1,2), 1, 3, byrow = TRUE))
det.sum.total <- as.data.frame(count(vdat[vdat$is_blank==0,], common_name))
det.sum.total <- det.sum.total[order(det.sum.total$n),]

par(mar=c(5,16,1,1))
barplot(det.sum.total$n, names.arg = paste0(det.sum.total$common_name, 
                                           " (n =", det.sum.total$n,")")   , las=1, cex.names=1, xlab="Total detections", horiz=T)
i <-1


for(i in 1:nrow(det.sum.total))
{
  tmp <- subset(vdat, common_name==det.sum.total$common_name[i])
  det.sum.total$Locations[i] <- length(unique(tmp$deployment_location_id))
}
par(mar=c(5,1,1,1))

barplot(det.sum.total$Locations/n.loc, las=1, cex.names=0.7, xlab="Proportion of sites detected", horiz=T, xlim=c(0,1))
abline(v=1, lty=2)

```

To date we have `r length(unique(vdat$common_name))` unique classifications. Of these, `r length(unique(vdat[vdat$species!="",]$sp))` are to full species level, `r length(unique(vdat[vdat$species=="",]$genus))` to genus level, the rest are higher (family/order). 

Note you can use the slider bar at the bottom to control the the x-axis extent. 

```{r, echo=F}
# MAke it in plotly?
#Order

det.sum.total$order <-1:nrow(det.sum.total)

yform <- list(categoryorder = "array",
              categoryarray = det.sum.total$common_name,
              tickfont = list(size = 7))

xform <- list(rangeslider = list(type = "n"))

fig <- plot_ly(x = det.sum.total$n, y = det.sum.total$common_name, type = 'bar', orientation = 'h',
               height=2000)%>% 
 layout(yaxis = yform, xaxis=xform)

fig

```

# Diel time
For any species detected more than 100 times, we will plot when they were detected:

```{r, echo=F, warning=F, message=F}
vdat$hours <- hour(vdat$timestamp) + minute(vdat$timestamp)/60 + second(vdat$timestamp)/(60*60)

# Susbet to more than 100 captures
tmp <- det.sum.total[det.sum.total$n>100,]

yform <- list(categoryorder = "array",
              categoryarray = tmp$common_name,
              tickfont = list(size = 7))
tmp2 <- vdat[vdat$common_name %in% yform$categoryarray,]

fig <- plot_ly(x = tmp2$hours, y = tmp2$common_name,type="scatter",
               height=1000, text=tmp2$deployment_id, hoverinfo='text',
               mode   = 'markers',
               marker = list(size = 5,
                             color = 'rgba(50, 100, 255, .2)',
                             line = list(color = 'rgba(0, 0, 0, 0)',
                                         width = 0))) %>% 
              layout(yaxis = yform)

fig

```


# Detection distances 

We have been recording detection distances for species at stations with sticks. Below is a summary of the detection distances so far.

```{r, echo=F, include=F}
# Data cleaning
table(dat$individual_animal_notes)
dat$individual_animal_notes[dat$individual_animal_notes=="0--2"] <- "0-2"
dat$individual_animal_notes[dat$individual_animal_notes=="0.2"] <- "0-2"
dat$individual_animal_notes[dat$individual_animal_notes=="0-4"] <- ""
dat$individual_animal_notes[dat$individual_animal_notes=="2.4"] <- "2-4"
dat$individual_animal_notes[dat$individual_animal_notes=="2--4"] <- "2-4"
dat$individual_animal_notes[dat$individual_animal_notes=="2-4"] <- "2-4"
dat$individual_animal_notes[dat$individual_animal_notes=="2;4"] <- "2-4"
dat$individual_animal_notes[dat$individual_animal_notes=="2=4"] <- "2-4"
dat$individual_animal_notes[dat$individual_animal_notes=="24"] <- "2-4"
dat$individual_animal_notes[dat$individual_animal_notes=="6-8"] <- "6+"
dat$individual_animal_notes[dat$individual_animal_notes=="6 plus"] <- "6+"
dat$individual_animal_notes[dat$individual_animal_notes=="+6"] <- "6+"
dat$individual_animal_notes[dat$individual_animal_notes=="6 +"] <- "6+"


dat_dist <- dat[dat$individual_animal_notes!="",]

# Remove ambiguous values
dat_dist <- dat_dist[dat_dist$individual_animal_notes!="0-4",]

kbl(table(dat_dist$individual_animal_notes))


dat_dist$individual_animal_notes <- as.factor(dat_dist$individual_animal_notes)

tmp <- unique(dat_dist$deployment_location_id)

```

So far we have data from `r length(tmp)` different locations, of which `r length(tmp[substr(tmp,1,3)=="LAC"])` are from on trail sites and `r length(tmp[substr(tmp,1,3)=="OFF"])` are from off trail sites.

```{r, echo=F, message=F, warning=F}

tmp <- dat_dist %>% group_by(common_name , individual_animal_notes, .drop=FALSE) %>% summarise("freq"=n()) 

# Remove boring classifications
#unique(tmp$common_name)
tmp <- tmp[!tmp$common_name %in% c("Animal", "Bird", "Cervidae Family", "Columbidae Family", "Dasyproctidae Family", "Lizards and Snakes", "Mammal", "No CV Result", "Penelope Species"), ]


par(mfrow=c(3,3))
par(mar=c(4,4,2,1))

for(i in 1:length(unique(tmp$common_name)))
{
   barplot(tmp[tmp$common_name==unique(tmp$common_name)[i],]$freq,las=1,
           main=unique(tmp$common_name)[i], names=levels(tmp$individual_animal_notes))
}


```

# Speed data

```{r, echo=F, warning=F, message=F, include=F}

#Calculate speed in kmph
library(lubridate)
str(speed)
speed$interval_sec <- as.numeric(difftime(speed$img2_timestamp,speed$img1_timestamp))
speed$kmph <- ((speed$dist_m/speed$interval_sec)/1000)*60*60

speed_sum <- speed %>% group_by(common_name) %>% summarize(mean=mean(kmph,na.rm=T), sd=sd(kmph,na.rm=T)) %>% arrange(mean)
speed_sum <- speed_sum[is.na(speed_sum$mean)==F,]
speed_sp <- speed_sum$common_name[is.na(speed_sum$mean)==F]
```

Make a plot of all the data:

```{r, echo=F}
par(mfrow=c(1,1))
par(mar=c(5,8,1,1))

plot(c(min(speed$kmph,na.rm=T), max(speed$kmph,na.rm=T)), c(1, length(speed_sp)), type="n", yaxt="n", ylab="", xlab="Speed kmph")
axis(2,at=1:length(speed_sp), labels=speed_sp, las=1, cex.axis=0.7)
i <- 1
for(i in 1:length(speed_sp))
{
  points(i~ speed_sum$mean[speed_sum$common_name==speed_sp[i]], pch=19, cex=1.5)
  tmp <- speed[speed$common_name==speed_sp[i] & is.na(speed$common_name)==F,]
  points(rep(i, nrow(tmp))~ tmp$kmph, pch=19, cex=0.7, col="darkgrey")
}
```

# Carcass patterns 

Detections in the first 10 days (240 hours) after carcasses placement.

```{r, echo=F, warning=F, message=F}
# I need a mechanism to record the exact time we left the carcass!


vul_dat <- vdat[substr(vdat$deployment_id,1,3)=="VUL",]
i <- 1
for(i in 1:length(unique(vul_dat$deployment_location_id)))
{
  par(mfrow=c(1,3))
  par(mar=c(10,4,1,1))
  
  
  tmp <- vul_dat[vul_dat$deployment_location_id==unique(vul_dat$deployment_location_id)[i],]
  tmp$timer_hrs <- as.numeric(difftime(tmp$timestamp, carcass$Date_time_deployed[carcass$deployment_location_id==unique(vul_dat$deployment_location_id)[i] & is.na(carcass$deployment_location_id)==F], units="hours"))
  tmp <-tmp[order(tmp$timer),]
  # Remove homosapians and blanks
  tmp <- tmp[tmp$sp!="Homo sapiens" & tmp$sp!=" ",]
  tmp <- tmp[tmp$timer_hrs<240,]
    #min(dep$start_date[dep$deployment_location_id==unique(vul_dat$deployment_location_id)[i]])
  hist(tmp$timer_hrs, main=tmp$deployment_location_id[1], breaks=seq(0,240, 2), las=1, xlab="Hours since deployment", ylab="Number of observations (2 hr bins")
  tmp_summary <- tmp %>% group_by(sp) %>% summarise(Obs=n(),Count=sum(number_of_objects),RAI=sum(number_of_objects)/n() ,Timing=mean(hours), Timing_var=sd(hours))
  # Barplot
  tmp_summary <- tmp_summary[order(tmp_summary$Obs),]
  barplot(tmp_summary$Obs, names.arg=tmp_summary$sp, las=2, cex.names=0.8, ylab="Number of observations")
  # # Activity
  # tmp_summary <- tmp_summary[order(tmp_summary$Timing),]
  # barplot(tmp_summary$Timing, names.arg=tmp_summary$sp, las=2, cex.names=0.8, ylab="Mean detection time")
  # Arrival order
  tmp2 <- tmp[duplicated(tmp$common_name)==F,]
  plot(c(1,3), c(1,nrow(tmp2)), xaxt="n", yaxt="n", type="n", xlab="", ylab="", main= "Arrival order")
  points(rep(1, times=nrow(tmp2)), nrow(tmp2):1, pch=19)
  text(rep(1.1, times=nrow(tmp2)),nrow(tmp2):1, paste(tmp2$common_name, "-", round(tmp2$timer_hrs,1), "h"), pos=4, cex=0.8 )
  axis(2, at=nrow(tmp2):1, labels=1:nrow(tmp2), las=1)
}


```


### When are different species detected

Not yet implemented.

```{r}
library(plotrix)




```



# Detection check

We also need to check if all of the deployments in Wildlife Insights have associated image data.

Of the deployments currently in Wildlife Insights, **`r paste0(round( (length(dep$deployment_id[dep$deployment_id %in% dat$deployment_id])/nrow(dep))*100,2), "%")`** of deployments have associated image data.

Check we don't have duplicates: `r length(dep$deployment_id[duplicated(dep$deployment_id)==TRUE])`

```{r, echo=F, eval=F}
#If that number is >0, what observations does it relate to?
dep[dep$deployment_id %in% dep$deployment_id[duplicated(dep$deployment_id)==TRUE],]
```


I have emailed Wildlife Insights about this. I didn't think it was possible. 

*Note - our current workflow should mean that deployments ALWAYS have image data. If the percentage is <100% then there is a problem*


```{r, echo=F}
#We can also check to see if any deployments need truncating - for example if the camera ceased to be active:
dat$sp <- paste(dat$genus, dat$species)


# Plot the start and end date for each deployment
#fig <- plot_ly(x = dep$start_date[1:5], y = dep$deployment_id[1:5],type="scatter",
#               height=2000)
plot.order <- dep$deployment_id[order(dep$deployment_id)]
# Remove duplicates
plot.order<- plot.order[duplicated(plot.order)==F]
#fig
tmp4 <- data.frame("deployment_id"=plot.order, "Plot.grp"=ceiling(1:length(plot.order)/20))

eff <- left_join(tmp4,dep,by="deployment_id")
eff <- eff[duplicated(eff)==FALSE,]


i<- 15
j <- 1
for(j in 1:length(unique(eff$Plot.grp)))
{
    layout(matrix(c(1,1,1,2), 1, 4, byrow = TRUE))
    par(mar=c(2,10,1,1))
    #str(eff)
    plot(c(min(eff$start_date, na.rm=T), max(eff$end_date, na.rm=T)),      c(1,length(unique(eff$deployment_id[eff$Plot.grp==j]))), las=1, ylab="", xlab="", type="n", yaxt="n")
    
    axis(2, at= length(unique(eff$deployment_id[eff$Plot.grp==j])):1, labels= unique(eff$deployment_id[eff$Plot.grp==j]), las=1, cex.axis=1)
    #mtext("Camera Deployment ID", 2, 4)
    # Make lines for each of the cameras
    for(i in length(eff$deployment_id[eff$Plot.grp==j]):1)
    {
      abline(h=i, col=rgb(0,0,0,0.1))
      tmp <- eff[eff$deployment_id==rev(eff$deployment_id[eff$Plot.grp==j])[i],]
      
      tmp2 <- dat[dat$deployment_id==tmp$deployment_id[1],]
      #tmp2 <- left_join(tmp2, tmp3, by = "sp")
      points(tmp2$timestamp, rep(i,nrow(tmp2)), pch="|", col= tmp2$is_blank+2)
    
      for(k in 1:nrow(tmp))
        {
          lines(c(tmp$start_date[k],
                           tmp$end_date[k]),
                c(i,i), lwd=2)
        }
      }
    par(mar=c(0,0,1,0))
    plot.new()
    legend("topleft", legend=c("Animal", "Blank"), fill=c(2,3), xpd=TRUE, cex=1.1 )

}


```




# Full species list

A full list of labelled species in alphabetic order:

```{r, echo=F}
#head(vdat)
tmp <- vdat[, c("class","order","family", "genus","species")]

tmp <- tmp[duplicated(tmp)==F,]
tmp <- tmp[ with(tmp, order(class,order,family, genus,species)),]
row.names(tmp) <- NULL

tmp %>%
  kbl() %>%
  kable_styling(full_width = T) %>%
  #column_spec(1, bold = T, border_right = T )%>%
  scroll_box( height = "800px")




```

# Meta data

Summary information for the project:

```{r, echo=F}
t(pro) %>%
  kbl() %>%
  kable_styling(full_width = T) %>%
  column_spec(1, bold = T, border_right = T)


```

# write important shapefiles

```{r}

write.csv(dat, "data/processed-data/LosAmigos_all_image_data.csv", row.names=F)

feats <- all.locs[all.locs$Project == "Features",]



feats_sf <- st_as_sf(feats, coords=c("Longitude", "Latitude"), crs=4326)

#st_write(feats_sf, "data/shapefiles/Feature_locations.shp", delete_layer=T)

```
